Moved UpdateExternalWorkProcessCommand model and validator into the Common.Application project so it can be called from outside of the Work service area.
