+++ Moda.Services/Moda.AppIntegration/src/Moda.AppIntegration.Application/Connections/Managers/AzureDevOpsBoardsInitManager.cs
@@ -7,7 +7,6 @@ using Moda.Common.Application.Requests.WorkManagement;
 using Moda.Common.Domain.Models;
 using Moda.Common.Models;
 using Moda.Work.Application.Workflows.Dtos;
-using Moda.Work.Application.WorkProcesses.Commands;
 using Moda.Work.Application.WorkProcesses.Queries;
 using Moda.Work.Application.Workspaces.Queries;
 using Moda.Work.Application.WorkStatuses.Commands;
